models/model1/requirements.txt
Flask
torch
Pillow
scikit-learn
matplotlib
scikit-image
flask-cors